requirements.txt
Flask
numpy
opencv-python
Pillow
torch
torchvision
timm
